add bottle-based Python backend that does a proxy to NodeJS
You should now start `python3 lexonomy.py`, it will autostart
the NodeJS backend as well if it is not running.

For the moment the only thing it does is serving static files,
anything else is redirected to NodeJS.

It can be deployed using the builtin single-threaded web server,
a multithreaded webserver from the "paste" package or (preferably)
as a simple CGI script in Apache or any other standalone web server.