Work on separating element definition ID and element name
Lexonomy ties the element name to the ruleset for that element (i.e. which attributes, text content, children it may contain). This is problematic for document structures where elements can have different meanings based on their position in the document. As is often the case in more complex document formats such as TEI(-lex0).

With this commit the element's name and its id/key in the xema are separated. The definition receives an extra property "elementName" which is used when rendering the element.

Though there may be different rulesets/ids for a certain element name now, an element may only specify one ruleset per element for its children.
So <cit> elements under a <translate> element will always have the same definition, but a <cit> element under a <entry> element may have a different definition.