Fix saving configs requiring re-index, and fix manual re-indexing
The button on the config page to trigger a re-index (should it be required) didn't work correctly.
Additionally, automatic re-indexing on saving a config that requires entries to be re-indexed threw a python error.
Re-indexing when saving the config is no longer required, as it can be done on a per-entry basis as they are accessed.