Always pass ID to createEntry to prevent entries getting duplicated sometimes
By default createEntry attempts to read the entry ID from the xml, but when that fails, it assumes (sometimes wrongly!) that the entry is completely new.
So when we have a specific ID we need to update, pass it to createEntry to ensure it overwrites THAT entry, even when the xml didn't contain a proper lxnm:id attribute at the root that we could extract.

Issue #289