Fix npm install throwing warnings
It was caused by rollup-plugin-uglify because it hasn't been updated in a long time and requests an outdated version of rollup, which conflicts with the more modern version requested by the other rollup plugins.
We were already using terser instead of uglify to perform minification, so it was safe to remove.

Also removed a few unused packages.